Unifying build process.
