+++ src/paging/mod.rs
@@ -1,13 +1,17 @@
 mod frame_alloc;
 mod multi_level;
+#[cfg(feature = "hypervisor")]
 mod multi_level_x4;
 mod page_table;
+#[cfg(feature = "hypervisor")]
 mod page_table_x4;
 mod recursive;
 
 pub use self::frame_alloc::*;
 pub use self::multi_level::*;
+#[cfg(feature = "hypervisor")]
 pub use self::multi_level_x4::*;
 pub use self::page_table::*;
+#[cfg(feature = "hypervisor")]
 pub use self::page_table_x4::*;
 pub use self::recursive::*;
